playwright: refactored to separate page object from app object
