+++ packages/example-v8/src/stories/Dev.spec.ts
@@ -1,9 +1,9 @@
 import {test} from "@playwright/test";
 import {assertStorybookIsRunning} from "../tests/utils";
-import StorybookPageObject, {type ControlExpectation} from "../tests/utils/StorybookPage";
 import {TEST_TIMEOUT_MS} from "../tests/utils/constants";
+import StorybookPageObject from "../tests/utils/StorybookPage";
+import type {ControlExpectation} from "../tests/utils/StoryPageObject";
 
-// todo colocate tests with source files
 test.beforeAll(assertStorybookIsRunning);
 
 test.beforeEach(async ({page}) => {
@@ -85,197 +85,209 @@ function createExpectedDefaultVisibleControls(): Record<string, ControlExpectati
 
 test("supports checking and unchecking root level boolean control", async ({page}) => {
   const storybookPage = new StorybookPageObject(page);
-  await storybookPage.assert.actualConfigMatches(createDefaultOutputConfig());
-  await storybookPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
+  await storybookPage.action.clickStoryById("stories-dev--enabled");
+
+  const storyPage = storybookPage.activeStoryPage;
+  await storyPage.assert.actualConfigMatches(createDefaultOutputConfig());
+  await storyPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
 
   const newConfig = createDefaultOutputConfig();
   const newVisibleControls = createExpectedDefaultVisibleControls();
 
   // uncheck control
-  const controlInput = storybookPage.getLocatorForControlInput("bool");
+  const controlInput = storyPage.getLocatorForControlInput("bool");
   await controlInput.click();
   newConfig.bool = false;
   newVisibleControls.bool = false;
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(newConfig);
-  await storybookPage.assert.controlsMatch(newVisibleControls);
+  await storyPage.assert.actualConfigMatches(newConfig);
+  await storyPage.assert.controlsMatch(newVisibleControls);
 
   // re-check control
   await controlInput.click();
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(createDefaultOutputConfig());
-  await storybookPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
+  await storyPage.assert.actualConfigMatches(createDefaultOutputConfig());
+  await storyPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
 
   // re-uncheck control
   await controlInput.click();
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(newConfig);
-  await storybookPage.assert.controlsMatch(newVisibleControls);
+  await storyPage.assert.actualConfigMatches(newConfig);
+  await storyPage.assert.controlsMatch(newVisibleControls);
 });
 
 test("supports checking and unchecking nested boolean control", async ({page}) => {
   const storybookPage = new StorybookPageObject(page);
-  await storybookPage.assert.actualConfigMatches(createDefaultOutputConfig());
-  await storybookPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
+  await storybookPage.action.clickStoryById("stories-dev--enabled");
+
+  const storyPage = storybookPage.activeStoryPage;
+  await storyPage.assert.actualConfigMatches(createDefaultOutputConfig());
+  await storyPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
 
   const newConfig = createDefaultOutputConfig();
   const newVisibleControls = createExpectedDefaultVisibleControls();
 
   // check control
-  const controlInput = storybookPage.getLocatorForControlInput("nested.bool");
+  const controlInput = storyPage.getLocatorForControlInput("nested.bool");
   await controlInput.click();
   newConfig.nested.bool = true;
   newVisibleControls["nested.bool"] = true;
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(newConfig);
-  await storybookPage.assert.controlsMatch(newVisibleControls);
+  await storyPage.assert.actualConfigMatches(newConfig);
+  await storyPage.assert.controlsMatch(newVisibleControls);
 
   // re-uncheck control
   await controlInput.click();
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(createDefaultOutputConfig());
-  await storybookPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
+  await storyPage.assert.actualConfigMatches(createDefaultOutputConfig());
+  await storyPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
 
   // re-check control
   await controlInput.click();
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(newConfig);
-  await storybookPage.assert.controlsMatch(newVisibleControls);
+  await storyPage.assert.actualConfigMatches(newConfig);
+  await storyPage.assert.controlsMatch(newVisibleControls);
 });
 
 test("supports checking and unchecking deep nested boolean control", async ({page}) => {
   const storybookPage = new StorybookPageObject(page);
-  await storybookPage.assert.actualConfigMatches(createDefaultOutputConfig());
-  await storybookPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
+  await storybookPage.action.clickStoryById("stories-dev--enabled");
+
+  const storyPage = storybookPage.activeStoryPage;
+  await storyPage.assert.actualConfigMatches(createDefaultOutputConfig());
+  await storyPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
 
   const expectedConfig = createDefaultOutputConfig();
   const expectedControls = createExpectedDefaultVisibleControls();
 
   // uncheck control
-  const controlInput = storybookPage.getLocatorForControlInput("nested.nested.bool");
+  const controlInput = storyPage.getLocatorForControlInput("nested.nested.bool");
   await controlInput.click();
   expectedConfig.nested.nested.bool = false;
   expectedControls["nested.nested.bool"] = false;
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(expectedConfig);
-  await storybookPage.assert.controlsMatch(expectedControls);
+  await storyPage.assert.actualConfigMatches(expectedConfig);
+  await storyPage.assert.controlsMatch(expectedControls);
 
   // re-check control
   await controlInput.click();
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(createDefaultOutputConfig());
-  await storybookPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
+  await storyPage.assert.actualConfigMatches(createDefaultOutputConfig());
+  await storyPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
 
   // re-uncheck control
   await controlInput.click();
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(expectedConfig);
-  await storybookPage.assert.controlsMatch(expectedControls);
+  await storyPage.assert.actualConfigMatches(expectedConfig);
+  await storyPage.assert.controlsMatch(expectedControls);
 });
 
 test("supports resetting controls", async ({page}) => {
   const storybookPage = new StorybookPageObject(page);
-  await storybookPage.assert.actualConfigMatches(createDefaultOutputConfig());
-  await storybookPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
+  await storybookPage.action.clickStoryById("stories-dev--enabled");
+
+  const storyPage = storybookPage.activeStoryPage;
+  await storyPage.assert.actualConfigMatches(createDefaultOutputConfig());
+  await storyPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
 
   const expectedConfig = createDefaultOutputConfig();
   const expectedControls = createExpectedDefaultVisibleControls();
 
   // uncheck root control
-  let controlInput = storybookPage.getLocatorForControlInput("bool");
+  let controlInput = storyPage.getLocatorForControlInput("bool");
   await controlInput.click();
   expectedConfig.bool = false;
   expectedControls.bool = false;
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(expectedConfig);
-  await storybookPage.assert.controlsMatch(expectedControls);
+  await storyPage.assert.actualConfigMatches(expectedConfig);
+  await storyPage.assert.controlsMatch(expectedControls);
 
   // change root string control
-  controlInput = storybookPage.getLocatorForControlInput("string");
+  controlInput = storyPage.getLocatorForControlInput("string");
   await controlInput.fill("new string");
   expectedConfig.string = "new string";
   expectedControls.string = "new string";
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(expectedConfig);
-  await storybookPage.assert.controlsMatch(expectedControls);
+  await storyPage.assert.actualConfigMatches(expectedConfig);
+  await storyPage.assert.controlsMatch(expectedControls);
 
   // change root number control
-  controlInput = storybookPage.getLocatorForControlInput("number");
+  controlInput = storyPage.getLocatorForControlInput("number");
   await controlInput.fill("905");
   expectedConfig.number = 905;
   expectedControls.number = 905;
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(expectedConfig);
-  await storybookPage.assert.controlsMatch(expectedControls);
+  await storyPage.assert.actualConfigMatches(expectedConfig);
+  await storyPage.assert.controlsMatch(expectedControls);
 
   // uncheck deeply nested boolean control
-  controlInput = storybookPage.getLocatorForControlInput("nested.nested.bool");
+  controlInput = storyPage.getLocatorForControlInput("nested.nested.bool");
   await controlInput.click();
   expectedConfig.nested.nested.bool = false;
   expectedControls["nested.nested.bool"] = false;
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(expectedConfig);
-  await storybookPage.assert.controlsMatch(expectedControls);
+  await storyPage.assert.actualConfigMatches(expectedConfig);
+  await storyPage.assert.controlsMatch(expectedControls);
 
   // change deeply nested string control
-  controlInput = storybookPage.getLocatorForControlInput("nested.nested.string");
+  controlInput = storyPage.getLocatorForControlInput("nested.nested.string");
   await controlInput.fill("new deeply nested string");
   expectedConfig.nested.nested.string = "new deeply nested string";
   expectedControls["nested.nested.string"] = "new deeply nested string";
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(expectedConfig);
-  await storybookPage.assert.controlsMatch(expectedControls);
+  await storyPage.assert.actualConfigMatches(expectedConfig);
+  await storyPage.assert.controlsMatch(expectedControls);
 
   // change deeply nested number control
-  controlInput = storybookPage.getLocatorForControlInput("nested.nested.number");
+  controlInput = storyPage.getLocatorForControlInput("nested.nested.number");
   await controlInput.fill("-99");
   expectedConfig.nested.nested.number = -99;
   expectedControls["nested.nested.number"] = -99;
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(expectedConfig);
-  await storybookPage.assert.controlsMatch(expectedControls);
+  await storyPage.assert.actualConfigMatches(expectedConfig);
+  await storyPage.assert.controlsMatch(expectedControls);
 
   // change deeply nested number control with infinity value (make sure we can restore this)
-  controlInput = storybookPage.getLocatorForControlInput("nested.nested.infinity");
+  controlInput = storyPage.getLocatorForControlInput("nested.nested.infinity");
   await controlInput.fill("-42");
   expectedConfig.nested.nested.infinity = -42;
   expectedControls["nested.nested.infinity"] = -42;
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(expectedConfig);
-  await storybookPage.assert.controlsMatch(expectedControls);
+  await storyPage.assert.actualConfigMatches(expectedConfig);
+  await storyPage.assert.controlsMatch(expectedControls);
 
   // change deeply nested number control with NaN value (make sure we can restore this)
-  controlInput = storybookPage.getLocatorForControlInput("nested.nested.NaNValue");
+  controlInput = storyPage.getLocatorForControlInput("nested.nested.NaNValue");
   await controlInput.fill("-82");
   expectedConfig.nested.nested.NaNValue = -82;
   expectedControls["nested.nested.NaNValue"] = -82;
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(expectedConfig);
-  await storybookPage.assert.controlsMatch(expectedControls);
+  await storyPage.assert.actualConfigMatches(expectedConfig);
+  await storyPage.assert.controlsMatch(expectedControls);
 
   // reset controls
-  await storybookPage.resetControlsButtonLocator.click();
+  await storyPage.resetControlsButtonLocator.click();
 
   // assert change
-  await storybookPage.assert.actualConfigMatches(createDefaultOutputConfig());
-  await storybookPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
+  await storyPage.assert.actualConfigMatches(createDefaultOutputConfig());
+  await storyPage.assert.controlsMatch(createExpectedDefaultVisibleControls());
 });
 
 // also tests it handles objects with existing properties partially defined by argTypes
@@ -284,7 +296,9 @@ test("supports customising existing property control with initial primitive valu
 }) => {
   const storybookPage = new StorybookPageObject(page);
   await storybookPage.action.clickStoryById("stories-dev--with-custom-controls");
-  await storybookPage.assert.controlsMatch({
+
+  const storyPage = storybookPage.activeStoryPage;
+  await storyPage.assert.controlsMatch({
     "someObject.anyString": "anyString",
     "someObject.enumString": {
       type: "radio",
@@ -294,7 +308,7 @@ test("supports customising existing property control with initial primitive valu
   });
 
   // initial value not affected by custom controls
-  await storybookPage.assert.actualConfigMatches({
+  await storyPage.assert.actualConfigMatches({
     someObject: {
       anyString: "anyString",
       enumString: "value2",
@@ -308,7 +322,9 @@ test("supports customising non-existing property control without initial value",
   await storybookPage.action.clickStoryById(
     "stories-dev--with-custom-controls-for-non-existing-property",
   );
-  await storybookPage.assert.controlsMatch({
+
+  const storyPage = storybookPage.activeStoryPage;
+  await storyPage.assert.controlsMatch({
     "someObject.anyString": "anyString",
     "someObject.enumString": "value2",
     "someObject.unknown": {
@@ -319,7 +335,7 @@ test("supports customising non-existing property control without initial value",
   });
 
   // initial value not affected by custom controls
-  await storybookPage.assert.actualConfigMatches({
+  await storyPage.assert.actualConfigMatches({
     someObject: {
       anyString: "anyString",
       enumString: "value2",
@@ -330,7 +346,9 @@ test("supports customising non-existing property control without initial value",
 test("supports control matchers", async ({page}) => {
   const storybookPage = new StorybookPageObject(page);
   await storybookPage.action.clickStoryById("stories-dev--with-control-matchers");
-  await storybookPage.assert.controlsMatch({
+
+  const storyPage = storybookPage.activeStoryPage;
+  await storyPage.assert.controlsMatch({
     // control matched correctly
     "color.color": {
       type: "color",
@@ -340,7 +358,7 @@ test("supports control matchers", async ({page}) => {
     "color.description": "Very red",
   });
 
-  await storybookPage.assert.actualConfigMatches({
+  await storyPage.assert.actualConfigMatches({
     color: {
       color: "#f00",
       description: "Very red",
@@ -352,12 +370,13 @@ test("shows empty object and array controls", async ({page}) => {
   const storybookPage = new StorybookPageObject(page);
   await storybookPage.action.clickStoryById("stories-dev--with-empty-initial-args");
 
-  await storybookPage.assert.controlsMatch({
+  const storyPage = storybookPage.activeStoryPage;
+  await storyPage.assert.controlsMatch({
     emptyObj: {type: "json", valueText: "{}"}, // empty object shown
     emptyArray: {type: "json", valueText: "[]"},
   });
 
-  await storybookPage.assert.actualConfigMatches({
+  await storyPage.assert.actualConfigMatches({
     emptyObj: {},
     emptyArray: [],
   });
@@ -367,7 +386,8 @@ test("handles object arg value overridden by argType", async ({page}) => {
   const storybookPage = new StorybookPageObject(page);
   await storybookPage.action.clickStoryById("stories-dev--with-overridden-object-arg");
 
-  await storybookPage.assert.controlsMatch({
+  const storyPage = storybookPage.activeStoryPage;
+  await storyPage.assert.controlsMatch({
     // control specified by arg type shown
     "someObject.obj2WithArgType": {
       type: "json",
@@ -378,7 +398,7 @@ test("handles object arg value overridden by argType", async ({page}) => {
     "someObject.obj1.bar1": "bar1",
   });
 
-  await storybookPage.assert.actualConfigMatches({
+  await storyPage.assert.actualConfigMatches({
     someObject: {
       obj1: {
         foo1: "foo1",
